# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaS Multi-Niche API",
        "description": "API RESTful pour système de réservation multi-niche (paragliding, surfing, diving, etc.)",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server (Développement)"
        },
        {
            "url": "https://api.example.com",
            "description": "API Server (Production)"
        }
    ],
    "paths": {
...
        "/api/v1/instructors": {
            "get": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Liste des instructeurs",
                "description": "Retourne la liste des instructeurs actifs, optionnellement filtrés par type d'activité",
                "operationId": "listInstructors",
                "parameters": [
                    {
                        "name": "activity_type",
                        "in": "query",
                        "description": "Filtrer par type d'activité (paragliding, surfing, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "paragliding"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des instructeurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Instructor"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "organization": []
                    }
                ]
            }
        },
        "/api/v1/instructors/by-activity/{activity_type}": {
            "get": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Instructeurs par type d'activité",
                "description": "Retourne la liste des instructeurs pour un type d'activité spécifique",
                "operationId": "instructorsByActivity",
                "parameters": [
                    {
                        "name": "activity_type",
                        "in": "path",
                        "description": "Type d'activité",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "paragliding"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des instructeurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Instructor"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "organization": []
                    }
                ]
            }
        },
        "/api/v1/instructors/{id}": {
            "get": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Détails d'un instructeur",
                "description": "Retourne les détails complets d'un instructeur (admin uniquement)",
                "operationId": "getInstructor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'instructeur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Instructor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Instructeur non trouvé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {
                        "organization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Instructor": {
                "title": "Instructor",
                "description": "Modèle d'instructeur",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "activity_types": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "paragliding",
                            "surfing"
                        ]
                    },
                    "license_number": {
                        "type": "string",
                        "nullable": true
                    },
                    "experience_years": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "max_sessions_per_day": {
                        "type": "integer",
                        "example": 8,
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "can_accept_instant_bookings": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Obtenez le token via /api/v1/auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "organization": {
                "type": "apiKey",
                "description": "ID de l'organisation (tenant). Requis pour l'isolation multi-tenant.",
                "name": "X-Organization-ID",
                "in": "header"
            }
        }
    },
...
}